+++ serialization/dotnet/json/Microsoft.Kiota.Serialization.Json.Tests/JsonParseNodeTests.cs
@@ -14,6 +14,7 @@ namespace Microsoft.Kiota.Serialization.Json.Tests
                                             "        \"+1 412 555 0109\"\r\n" +
                                             "    ],\r\n" +
                                             "    \"displayName\": \"Megan Bowen\",\r\n" +
+                                            "    \"numbers\":\"one,two,thirtytwo\"," +
                                             "    \"givenName\": \"Megan\",\r\n" +
                                             "    \"accountEnabled\": true,\r\n" +
                                             "    \"createdDateTime\": \"2017 -07-29T03:07:25Z\",\r\n" +
@@ -45,6 +46,7 @@ namespace Microsoft.Kiota.Serialization.Json.Tests
             Assert.True(testEntity.AdditionalData.ContainsKey("mobilePhone"));
             Assert.Equal("Auditor", testEntity.AdditionalData["jobTitle"]);
             Assert.Equal("48d31887-5fad-4d73-a9f5-3c356e68a038", testEntity.Id);
+            Assert.Equal(TestEnum.One | TestEnum.Two, testEntity.Numbers ); // Unknown enum value is not included
         }
 
         [Fact]